requirements.txt
pyserial
PySide6
pyside6-helpers @ git@github.com:MrFrangipane/pyside6-helpers.git@main
python-arduino-serial @ git@github.com:MrFrangipane/python-arduino-serial.git@main
